+++ hbw/app/models/hbw/activiti/adapter.rb
@@ -25,12 +25,12 @@ module HBW
         response.status == 200
       end
 
-      def get_variables(user, entity_class, entity_code)
+      def get_variables(user, entity_class, entity_code, initial_variables)
         [
           { name: :initiator,                    value: user.id,     type: :string },
           { name: :initiatorEmail,               value: user.email,  type: :string },
           { name: entity_code_key(entity_class), value: entity_code, type: :string }
-        ]
+        ].concat(initial_variables)
       end
 
       def process_definition_for_key_like(key)